+++ src/components/Heading.mjs
@@ -1,18 +1,27 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import { stylesHeading } from '../styles/stylesHeading'
+import { classNameProp } from '../utils/classNameProp'
 import { LinkText } from './LinkText'
 
-export const Heading = React.forwardRef(({ id, children, ...props }, ref) => (
-  <h1 className={stylesHeading.className} id={id} {...props} ref={ref}>
-    {id ? <LinkText href={`#${id}`}>{children}</LinkText> : children}
-    {stylesHeading.styles}
-  </h1>
-))
+export const Heading = React.forwardRef(
+  ({ id, className, children, ...props }, ref) => (
+    <h1
+      className={classNameProp(stylesHeading.className, className)}
+      id={id}
+      {...props}
+      ref={ref}
+    >
+      {id ? <LinkText href={`#${id}`}>{children}</LinkText> : children}
+      {stylesHeading.styles}
+    </h1>
+  )
+)
 
 Heading.displayName = 'Heading'
 
 Heading.propTypes = {
   id: PropTypes.string,
+  className: PropTypes.string,
   children: PropTypes.node.isRequired
 }