+++ src/components/LinkNav.mjs
@@ -1,10 +1,16 @@
 import PropTypes from 'prop-types'
 import React from 'react'
+import { classNameProp } from '../utils/classNameProp'
+import { propTypeChildren } from '../utils/propTypeChildren'
 
 export const LinkNav = React.forwardRef(
-  ({ active, children, ...props }, ref) => (
-    <a className={active ? 'active' : undefined} {...props} ref={ref}>
-      {children}
+  ({ active, className, ...props }, ref) => (
+    <>
+      <a
+        className={classNameProp(className, active && 'active')}
+        {...props}
+        ref={ref}
+      />
       <style jsx>{`
         a {
           display: flex;
@@ -42,7 +48,7 @@ export const LinkNav = React.forwardRef(
           );
         }
       `}</style>
-    </a>
+    </>
   )
 )
 
@@ -50,5 +56,6 @@ LinkNav.displayName = 'LinkNav'
 
 LinkNav.propTypes = {
   active: PropTypes.bool,
-  children: PropTypes.node.isRequired
+  className: PropTypes.string,
+  children: propTypeChildren.isRequired
 }