+++ src/components/LinkText.mjs
@@ -1,25 +1,23 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import { stylesLinkText } from '../styles/stylesLinkText'
+import { classNameProp } from '../utils/classNameProp'
+import { propTypeChildren } from '../utils/propTypeChildren'
 
-export const LinkText = React.forwardRef(
-  ({ className, children, ...props }, ref) => (
+export const LinkText = React.forwardRef(({ className, ...props }, ref) => (
+  <>
     <a
-      className={`${stylesLinkText.className}${
-        className ? ` ${className}` : ''
-      }`}
+      className={classNameProp(stylesLinkText.className, className)}
       {...props}
       ref={ref}
-    >
-      {children}
-      {stylesLinkText.styles}
-    </a>
-  )
-)
+    />
+    {stylesLinkText.styles}
+  </>
+))
 
 LinkText.displayName = 'LinkText'
 
 LinkText.propTypes = {
   className: PropTypes.string,
-  children: PropTypes.node.isRequired
+  children: propTypeChildren.isRequired
 }