+++ src/components/Para.mjs
@@ -1,23 +1,23 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import { stylesPara } from '../styles/stylesPara'
+import { classNameProp } from '../utils/classNameProp'
+import { propTypeChildren } from '../utils/propTypeChildren'
 
-export const Para = React.forwardRef(
-  ({ className, children, ...props }, ref) => (
+export const Para = React.forwardRef(({ className, ...props }, ref) => (
+  <>
     <p
-      className={`${stylesPara.className}${className ? ` ${className}` : ''}`}
+      className={classNameProp(stylesPara.className, className)}
       {...props}
       ref={ref}
-    >
-      {children}
-      {stylesPara.styles}
-    </p>
-  )
-)
+    />
+    {stylesPara.styles}
+  </>
+))
 
 Para.displayName = 'Para'
 
 Para.propTypes = {
   className: PropTypes.string,
-  children: PropTypes.node.isRequired
+  children: propTypeChildren.isRequired
 }