+++ src/index.mjs
@@ -1,3 +1,4 @@
 export * from './components'
 export * from './hooks'
 export * from './styles'
+export * from './utils'