+++ src/utils/index.mjs
@@ -1,0 +1,2 @@
+export { classNameProp } from './classNameProp'
+export { propTypeChildren } from './propTypeChildren'